Add support for array variables to the "expressions" module and enhance integration tests.
This means that array variables are now properly expanded for "has_any" and "has_all" methods, as well as using crunch's "all_missing" and "all_valid" comparison function when the target variable for "missing" and "valid" is an array.

Also, expand the test suite and the integration test script to cover all the expression features available for use in "exclusion filters".